Refactor: abstract into function
and clean hard coded to loop
